+++ GiftDecline/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Feature: [#2] Add option to cap gift taste reduction
 - Feature: [#3] Add option to delay gift taste reduction
+- Fix: Error when giving a quest item
 - Fix: Use configuration of the host in multiplayer
 - Fix: Clear gift taste changes when exiting to the title screen